Serve bull board with fastify and update server index module
