+++ package.json
@@ -16,7 +16,6 @@
     "dev-workers": "node --import tsx --watch src/startWorkers.ts",
     "dev-board": "node -r dotenv/config --import tsx --watch src/index.ts",
     "dev-api": "node --import tsx --watch src/index.ts --api-only",
-    "new-api": "node --import tsx --watch src/server.ts --api-only",
     "dev": "concurrently \"npm run dev-board\" \"npm run dev-workers\"",
     "import": "node --import=tsx scripts/import-spreadsheet-companies.ts",
     "update-report-url": "node --import=tsx scripts/update-spreadsheet-companies-urls.ts",
@@ -29,11 +28,10 @@
   "license": "MIT License",
   "dependencies": {
     "@asteasolutions/zod-to-openapi": "^7.3.0",
-    "@bull-board/api": "^6.5.3",
-    "@bull-board/express": "^6.5.3",
+    "@bull-board/api": "^6.6.1",
+    "@bull-board/fastify": "^6.6.1",
     "@fastify/swagger": "^9.4.1",
     "@prisma/client": "^5.22.0",
-    "@scalar/express-api-reference": "^0.2.75",
     "@scalar/fastify-api-reference": "^1.25.103",
     "@types/node": "^22.10.0",
     "bullmq": "^5.28.0",
@@ -49,11 +47,9 @@
     "jest": "^29.7.0",
     "openai": "^4.73.1",
     "pdf2pic": "^3.1.3",
-    "pino-http": "^10.3.0",
+    "pino-http": "^10.4.0",
     "prisma-zod-generator": "^0.8.13",
     "sharp": "^0.33.5",
-    "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.1",
     "tsx": "^4.19.2",
     "wikibase-sdk": "^10.2.1",
     "zod": "^3.23.8"
@@ -61,8 +57,8 @@
   "devDependencies": {
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
-    "@types/node": "^22.8.4",
-    "concurrently": "^9.1.0",
+    "@types/node": "^22.10.6",
+    "concurrently": "^9.1.2",
     "deepl-node": "^1.15.0",
     "exceljs": "^4.4.0",
     "jest": "^29.7.0",
@@ -70,7 +66,7 @@
     "prisma": "^5.22.0",
     "supertest": "^7.0.0",
     "ts-jest": "^29.2.5",
-    "typescript": "^5.7.2"
+    "typescript": "^5.7.3"
   },
   "overrides": {
     "uuid": "^9.0.1"
