+++ src/bull-board.ts
@@ -1,15 +1,17 @@
 import { createBullBoard } from '@bull-board/api'
 import { BullMQAdapter } from '@bull-board/api/bullMQAdapter'
-import { ExpressAdapter } from '@bull-board/express'
+import { FastifyAdapter } from '@bull-board/fastify'
+
 import { workers } from './workers'
+import apiConfig from './config/api'
 
 const queues = workers.map((worker) => worker.queue)
-const serverAdapter = new ExpressAdapter()
-serverAdapter.setBasePath('/admin/queues')
+const serverAdapter = new FastifyAdapter()
+serverAdapter.setBasePath(apiConfig.bullBoardBasePath)
 
 createBullBoard({
   queues: queues.map((queue) => new BullMQAdapter(queue)),
-  serverAdapter: serverAdapter,
+  serverAdapter,
   options: {
     uiConfig: {
       boardTitle: 'Klimatkollen',
@@ -17,4 +19,4 @@ createBullBoard({
   },
 })
 
-export default serverAdapter.getRouter()
+export default serverAdapter.registerPlugin()
